bgpd: fix logic for unset SUBGRP_STATUS_FORCE_UPDATES flag
Currently the SUBGRP_STATUS_FORCE_UPDATES flag is cleared in
bgp_generate_updgrp_packets(). It turns out that's not the right
place for doing it. Several topotests such as bgp_as_allow_in and
bgp_set_aspath_replace are failing due to premature clearing of
the flag.

In this patch the logic is moved to subgroup_announce_route(), and
the flag is cleared only after the whole table has been announced.

Signed-off-by: Enke Chen <enchen@paloaltonetworks.com>